Crate flowsamples

flowsamples is a set of sample flows, used to test and demonstrate flow, and different semantics and characteristics of flows that can be written.

Each subdirectory holds a self-contained flow sample, with flow definition, docs etc and some of them provide their own function implementations.

At the top-level there is a build.rs build script that iterates over all the sub-folders found and compiles the flow. That will include flow compilation to produce a manifest.json flow manifest. If any flow has provided function implementations then they will be compiled to WASM files. When the build has completed, all samples should be ready to be ran, using the flowr flow runner.

This flowsamples binary provides a way for the user to run a sample manually and for automated test of all samples.

See main for details on running the flow samples yourself directly.

If cargo test is run on this crate, then all the samples will be ran using the provided test input and test args (see names of test files below) and the Stdout and File output will be compared to a set of predefined Stdout and File output to determine if the sample ran correctly. No Stderr output is expected so if any is detected the sample is deemed to have not ran correctly.

Constants

TEST_ARGS_FILENAME 🔒

Name of file where flow arguments for a flow sample test are read from

TEST_FILE_FILENAME 🔒

Name of file used for file output of a sample

TEST_STDERR_FILENAME 🔒

Name of file where any Stderr will be written from while executing a flowsample in test mode

TEST_STDIN_FILENAME 🔒

Name of file where any Stdin will be read from while executing a flowsample in test mode

TEST_STDOUT_FILENAME 🔒

Name of file where any Stdout will be written while executing a flowsample in test mode

Functions

args 🔒

Read the flow args from a file and return them as a Vector of Strings that will be passed to flowr

main 🔒

Run one or all of the flowsamples by typing flowsamples or cargo run -p flowsamples at the command line

run_sample 🔒

Run one specific flow sample